+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0 : 2020-10-07
 
 - Initial release
+
+## 0.2.0 : 2020-04-12
+
+- 🥝 Implemented interactive shell session.
+- 🥝 Implemented support for fabric script execution.
+- 🥝 Implemented connection and integration with praetorian-api.